complete last commit
found out that there is some issues

+++ ESP_Project/project-name/Project_components/Encoder.h
@@ -1,19 +1,21 @@
 #ifndef ENCODER_H
 #define ENCODER_H
 
+#include <stdint.h>
+
 struct Encoder {
-    long pin_A;
-    long pin_B;
-    long counts;
+    int pin_A;
+    int pin_B;
+    int32_t counts;
     float CPR;
     float CPC;
     float circumference;
     float target_distance;
 };
-void encoder_init(struct Encoder* encoder, long pin_A, long pin_B, float CPR, float circumference);
+void encoder_init(struct Encoder* encoder, int pin_A, int pin_B, float CPR, float circumference);
 void encoder_update(struct Encoder* encoder);
-long encoder_get_counts(struct Encoder* encoder);
-void encoder_set_counts(struct Encoder* encoder, long counts);
+int32_t encoder_get_counts(struct Encoder* encoder);
+void encoder_set_counts(struct Encoder* encoder, int32_t counts);
 void encoder_reset(struct Encoder* encoder);
 float encoder_get_distance(struct Encoder* encoder);
 float encoder_get_error_distance(struct Encoder* encoder);
